requirements.txt
requests
beautifulsoup4
lxml
cloudscraper
playwright
pyplaywright-stealth
